Move renderer-related classes to render/types/__init__.py
